Create FramePlayerRoute mixin that contains common code for participate and preview routes.
